+++ ewatercycle/parametersetdb/datafiles.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import os
 import subprocess
 from abc import abstractmethod, ABC
 
@@ -29,9 +30,19 @@ class SubversionCopier(AbstractCopier):
     """Uses subversion export to copy files from source to target
     """
     def save(self, target):
+        if os.path.exists(target):
+            raise Exception('Target directory already exists, will not overwrite')
         subprocess.check_call(['svn', 'export', self.source, target])
 
 
+class SymlinkCopier(AbstractCopier):
+    """Creates symlink from source to target
+    """
+    def save(self, target):
+        os.symlink(self.source, target)
+
+
 DATAFILES_FORMATS = {
-    'svn': SubversionCopier
+    'svn': SubversionCopier,
+    'symlink': SymlinkCopier,
 }